SA_RAG/python/pyproject.toml
[project]
name = "sa-rag"
version = "1.0.0"
description = "SA-RAG Python Orchestrator"
license = {text = "MIT"}
requires-python = ">=3.8"
dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "python-multipart>=0.0.6",
    "pymupdf>=1.23.0",
    "python-docx>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=5.0.0",
]
